# openapi/swagger.yaml
swagger: '2.0'
info:
  description: Wishox API Dobcumentation; It's WIP
  version: 0.0.0
  title: Wishbox
host: api.gameparty.ir
basePath: /api/V0.0.0
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: access
...
schemes:
  - https
  - http
paths:
...
  /storage/upload/:
    post:
      tags:
        - Storage
      summary: upload file
      description: ''
      operationId: uploadFile
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - ApiKeyAuth: []
      parameters:
        - in: formData
          name: file
          type: file
          required: true
          description: Upload File
      responses:
        '201':
          description: successful operation
          schema:
            type: object
            properties:
              ok:
                type: boolean
                default: true
              data:
                type: object
                properties:
                  media_id:
                    type: string
              status:
                type: number
                default: 201
        '400':
          description: Bad Request
        '429':
          description: Throttled
...